move eval group to interpret_group file
